fix(SearchBar): demo拆解与规范
