+++ src/packages/searchbar/demo.tsx
@@ -1,31 +1,16 @@
-import React, { useState } from 'react'
-import {
-  ArrowLeft,
-  Photograph,
-  ArrowDown,
-  More,
-  Close,
-} from '@nutui/icons-react'
-import { SearchBar } from './searchbar'
-import Toast from '../toast'
+import React from 'react'
 import { useTranslate } from '../../sites/assets/locale'
-import ConfigProvider from '../configprovider'
-import PopOver from '../popover'
 
-type TSearchDemo = {
-  basePlaceholder: string
-  text: string
-  test: string
-  title1: string
-  title2: string
-  title3: string
-  title4: string
-  title5: string
-  title6: string
-  title7: string
-}
+import Demo1 from './demos/h5/demo1'
+import Demo2 from './demos/h5/demo2'
+import Demo3 from './demos/h5/demo3'
+import Demo4 from './demos/h5/demo4'
+import Demo5 from './demos/h5/demo5'
+import Demo6 from './demos/h5/demo6'
+import Demo7 from './demos/h5/demo7'
+
 const SearchBarDemo = () => {
-  const [translated] = useTranslate<TSearchDemo>({
+  const [translated] = useTranslate({
     'zh-CN': {
       basePlaceholder: '上京东，购好物',
       text: '文本',
@@ -75,98 +60,24 @@ const SearchBarDemo = () => {
       title7: 'pengaturan suai',
     },
   })
-  const itemList = [
-    { name: 'option1' },
-    { name: 'option2' },
-    { name: 'option3' },
-  ]
-  const [lightTheme, setLightTheme] = useState(false)
 
-  const [value, setValue] = useState('')
-  const change = (val: string, e: React.ChangeEvent<HTMLInputElement>) => {
-    setValue(val)
-  }
   return (
     <>
       <div className="demo">
         <h2>{translated.title1}</h2>
-        <SearchBar backable placeholder={translated.basePlaceholder} />
+        <Demo1 />
         <h2>{translated.title2}</h2>
-        <SearchBar shape="round" maxLength={5} />
+        <Demo2 />
         <h2>{translated.title3}</h2>
-        <ConfigProvider
-          theme={{
-            nutuiSearchbarBackground: 'var(--nutui-color-primary)',
-            nutuiSearchbarInputBackground: '#eee',
-            nutuiSearchbarInputTextAlign: 'right',
-          }}
-        >
-          <SearchBar onSearch={(value) => Toast.show(value)} />
-        </ConfigProvider>
+        <Demo3 />
         <h2>{translated.title4}</h2>
-        <SearchBar
-          left={translated.text}
-          right={translated.test}
-          onSearch={(value) => Toast.show(value)}
-        />
+        <Demo4 />
         <h2>{translated.title5}</h2>
-        <SearchBar
-          left={
-            <>
-              <ArrowLeft width={16} height={16} />
-              <Close width={16} height={16} />
-            </>
-          }
-          right={
-            <>
-              <span>{translated.test}</span>
-              <More width={16} height={16} />
-            </>
-          }
-          rightIn={
-            <Photograph
-              width={16}
-              height={16}
-              onClick={() => {
-                console.log('Photograph right in')
-              }}
-            />
-          }
-        />
+        <Demo5 />
         <h2>{translated.title7}</h2>
-        <SearchBar
-          leftIn={
-            <PopOver
-              visible={lightTheme}
-              onClick={() => {
-                lightTheme ? setLightTheme(false) : setLightTheme(true)
-              }}
-              list={itemList}
-            >
-              <div style={{ fontSize: '12px', width: '50px', display: 'flex' }}>
-                更多
-                <ArrowDown />
-              </div>
-            </PopOver>
-          }
-        />
+        <Demo6 />
         <h2>{translated.title6}</h2>
-        <SearchBar
-          onChange={(val: string, e: React.ChangeEvent<HTMLInputElement>) =>
-            change(val, e)
-          }
-          maxLength={10}
-        />
-        <div
-          style={{
-            height: '40px',
-            lineHeight: '40px',
-            color: '#666',
-            fontSize: '14px',
-          }}
-        >
-          {value}
-        </div>
+        <Demo7 />
       </div>
     </>
   )
