+++ src/packages/searchbar/demos/h5/demo2.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import { SearchBar } from '../../searchbar'
+
+const Demo2 = () => {
+  return <>
+    <SearchBar shape="round" maxLength={5} />
+  </>
+}
+export default Demo2;
