+++ src/packages/searchbar/demos/h5/demo3.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+import { SearchBar } from '../../searchbar'
+import ConfigProvider from '../../../configprovider'
+import Toast from '../../../toast';
+
+const Demo3 = () => {
+  return <>
+    <ConfigProvider
+      theme={{
+        nutuiSearchbarBackground: 'var(--nutui-color-primary)',
+        nutuiSearchbarInputBackground: '#eee',
+        nutuiSearchbarInputTextAlign: 'right',
+      }}
+    >
+      <SearchBar onSearch={(value) => Toast.show(value)} />
+    </ConfigProvider>
+  </>
+}
+export default Demo3;
