+++ src/packages/searchbar/demos/taro/demo1.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import { SearchBar } from '@nutui/nutui-react-taro';
+
+const Demo1 = () => {
+  return <>
+    <SearchBar backable placeholder="上京东，购好物" />
+  </>
+}
+export default Demo1;
