+++ src/packages/searchbar/demos/taro/demo2.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import { SearchBar } from '@nutui/nutui-react-taro';
+
+const Demo2 = () => {
+  return <>
+    <SearchBar shape="round" maxLength={5} />
+  </>
+}
+export default Demo2;
