+++ src/packages/searchbar/demos/taro/demo3.tsx
@@ -1,0 +1,28 @@
+import React, { useState } from 'react';
+import { SearchBar, ConfigProvider, Toast } from '@nutui/nutui-react-taro';
+const [show, SetShow] = useState(false)
+const toastShow = () => {
+  SetShow(true)
+}
+const Demo3 = () => {
+  return <>
+    <ConfigProvider
+      theme={{
+        nutuiSearchbarBackground: 'var(--nutui-color-primary)',
+        nutuiSearchbarInputBackground: '#eee',
+        nutuiSearchbarInputTextAlign: 'right',
+      }}
+    >
+      <SearchBar onSearch={() => toastShow()} />
+    </ConfigProvider>
+    <Toast
+      type="text"
+      visible={show}
+      content="search callback"
+      onClose={() => {
+        SetShow(false)
+      }}
+    />
+  </>
+}
+export default Demo3;
